Modify sample_from_wavefunction to work in all cases.
The amplitude vector for the pyquil Wavefunction can be an ndarray (with a few different shapes) or a Python list. In some cases, the current `sample_from_wavefunction` function doesn't work and returns the all 0 string for all samples (i.e., it fails silently).

With these changes, `sample_from_wavefunction` works if the pyquil Wavefunction is created with a column vector (in the form of an (X,1) ndarray), an np.array, and a Python list.

More importantly, these changes no longer reverse the bitstring when sampled.